Module lang_util::error

Error type definitions

Structs

KindDisplay

ListDisplay

OptionalSourceNumber
Display wrapper for optional source numbers in resolved positions

ParseError
A parsing error wrapped from lalrpop_util’s error type

ResolvedPosition
A position in the input stream that has been resolved into the more user-friendly line and column indices

TokenDescription
Information about a lexed token

TokenDescriptor
Information about a known token

Enums

ParseErrorKind
Parsing error kind

Traits

LexicalError
An error produced by lexical analysis

Token
Trait to implement for a token to be used with lang_util’s infrastructure

TokenExt
Methods provided for all Token implementations