Module lang_util::located

Located type definition

Structs

Located

Wraps an object with location data

LocatedBuilder

Builder for a Located struct

Enums

FileOverride

Represents a file location override

Traits

FileIdResolver

Trait for resolving file identifiers to file names

HasFileNumber

Trait for objects that can return the current file number

Resolver

Trait for objects that can resolve offsets to line and column numbers