Struct lang_util::located::LocatedBuilder

pub struct LocatedBuilder {
    pos: TextRange,
    current_file: Option<FileId>,
    path: Option<PathBuf>,
    file_override: FileOverride,
    line_number: u32,
    column: u32,
}

Builder for a Located struct

Fields
pos: TextRange

Position at which the error occurred

current_file: Option<FileId>

File identifier for the error

path: Option<PathBuf>

Path corresponding to the file identifier

file_override: FileOverride

Overriden file location

line_number: u32

Resolved line number

column: u32

Resolved column number

Implementations

impl LocatedBuilder

pub fn new() -> Self

Create a new builder for Located with default values

pub fn pos(self, pos: impl Into<TextRange>) -> Self

Set the raw position

pub fn current_file(self, file: impl Into<FileId>) -> Self

Set the file identifier

pub fn path(self, path: impl Into<PathBuf>) -> Self

Set the source path

pub fn file_override(self, file_override: impl Into<FileOverride>) -> Self

Set the source file override

pub fn line_number(self, line_number: u32) -> Self

Set the resolved line number

pub fn column(self, column: u32) -> Self

Set the resolved column number

pub fn resolve(self, resolver: &impl Resolver) -> Self

Resolve the raw offset (see LocatedBuilder::pos) to line and column information

pub fn resolve_file(self, resolver: &(impl Resolver + HasFileNumber)) -> Self

Resolve the raw offset (see LocatedBuilder::pos) to line and column information, and set the current file information

pub fn resolve_path(self, resolver: &impl FileIdResolver) -> Self

Resolve the given file id into a path name, and store it in this builder

pub fn finish<E>(self, inner: E) -> Located<E>

Build the final Located object from the given inner object