Struct lang_util::node::content::Node

pub struct Node<T: NodeContent> {
    pub content: T,
    pub span: Option<NodeSpan>,
}

A syntax node with span information

Fields
content: T

Contents of this syntax node

span: Option<NodeSpan>

Span in the input this node was parsed from

Implementations

impl<T: NodeContent> Node<T>

pub fn new(content: T, span: Option<NodeSpan>) -> Self

Create a new syntax node with span information

pub fn into_inner(self) -> T

Return the wrapped syntax node, discarding the span information

pub fn map<U: NodeContent>(self, f: impl FnOnce(T) -> U) -> Node<U>

Map this content of this node into a new node

pub fn map_spanned<U: NodeContent>(self, f: impl FnOnce(Self) -> U) -> Node<U>

Map this content of this node into a new node with the same span