Trait lang_util::token::Token

pub trait Token: Display {
    fn variant_name(&self) -> &'static str;

    fn parser_token(&self) -> &'static str;

    fn kinds(&self) -> &'static [&'static str];

    fn all_tokens() -> &'static [TokenDescriptor];
}

Trait to implement for a token to be used with lang_util’s infrastructure

Required methods

fn variant_name(&self) -> &'static str

Return the variant name of the current token

fn parser_token(&self) -> &'static str

Return the name used by the lalrpop parser for this token

fn kinds(&self) -> &'static [&'static str]

Return the token kinds this token belongs to

fn all_tokens() -> &'static [TokenDescriptor]

Return the descriptions for all known tokens